fix quote-preservation in QUERY_STRING
QUERY_STRING is supposed to be faithful to the transmitted value. If one 
requests e.g. /?!%21, the QUERY_STRING should preserve both the unquoted 
and the quoted exclamation mark. Doing otherwise breaks digest
authentication in e.g. a WSGI middleware.

We cannot use yarl here as preserving the original representation is a 
non-goal for yarl. We must resort to re-parsing the raw_path.

Link: https://github.com/aio-libs/yarl/issues/498
Signed-off-by: Helmut Grohne <helmut.grohne@intenta.de>